Complete first draft of landing page.
Completed the html and CSS for the first draft of my first landing page.
Is relatively polished but needs some work to make the page scale
consistently with the size of the page. I was able to make the
individual sections scale ok, but the page as a whole does not look good
when zooming out of shortening the page width.